Fixed bug that would sometimes lead to floating sample rates being used

+++ utime/io/psg/psg_extractors.py
@@ -38,7 +38,7 @@ def extract_from_edf(psg_file_path, header, include_channels, exclude_channels, 
                            stim_channel=None, verbose=False,
                            exclude=exclude_channels_short)
         # Update header with actually used sample rate
-        header["sample_rate"] = float(edf.info['sfreq'])
+        header["sample_rate"] = int(edf.info['sfreq'])
         return edf.get_data().T
 
 
@@ -52,7 +52,7 @@ def extract_from_wfdb(wfdb_file_path, include_channels, header, **kwargs):
     from wfdb.io import rdrecord
     rec = rdrecord(record_name=os.path.splitext(wfdb_file_path)[0],
                    channel_names=include_channels)
-    header["sample_rate"] = float(rec.fs)
+    header["sample_rate"] = int(rec.fs)
     return rec.p_signal
